+++ packages/pluggableWidgets/datagrid-web/src/components/WidgetRoot.tsx
@@ -1,5 +1,5 @@
 import classNames from "classnames";
-import { ReactElement, createElement } from "react";
+import { ReactElement, createElement, useRef, useMemo } from "react";
 import { SelectionMethod } from "../helpers/SelectActionHelper";
 
 type P = JSX.IntrinsicElements["div"];
@@ -12,10 +12,21 @@ export interface WidgetRootProps extends P {
 }
 
 export function WidgetRoot(props: WidgetRootProps): ReactElement {
+    const ref = useRef<HTMLDivElement>(null);
     const { className, selectionMethod, selection, exporting, children, ...rest } = props;
+    const style = useMemo(() => {
+        const s = { ...props.style };
+        if (exporting && ref.current) {
+            s.height = ref.current.offsetHeight;
+        }
+        return s;
+    }, [props.style, exporting]);
+
     return (
         <div
             {...rest}
+            ref={ref}
+            style={style}
             className={classNames(className, "widget-datagrid", {
                 "widget-datagrid-exporting": exporting,
                 "widget-datagrid-selectable-rows": selection,